accomodate the authenticator/validators split in interfaces and unit tests
Some of the unit tests were not in fact testing separate units but would
instead use the full authenticator. The authenticator/validators split
revealed this flaw.

The unit tests affected by the above were modified in order to both
actually test the separate units, and to avoid cyclic dependencies
between the authenticator and the validators. The TestAuthenticateTokenTimeout
test was almost fully rewritten because of it. The missing authenticator
test cases will be brought back in a new test in an upcoming commit
that also modifies the authenticator to be usable for the
TokenReviews REST storage.

The interfaces.go file was split between the authenticator/validators
so that the interfaces appear in their place of use, and so that there
are no other cyclic dependencies.