add a wrapper to upstream TokenReviews storage to the oauth apiserver
This wrapper allows us to configure the kubernetes API server to
forward token validation requests (TokenReviews) to this API server.

The wrapper uses the Create() method of the upstream's tokenreview.REST
to handle incoming TokenReview requests as the kubernetes API server
would have done itself.

The token authenticator extracted from our kubernetes API server
patches is put in place for the wrapped storage so that it is used
to validate the tokens from the incoming TokenReview requests.